pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plotseamap"
version = "0.1.0"
description = "Automatisierte Erstellung von Seegebietskarten mit OSM‑ und AIS‑Daten"
readme = "README.md"
authors = [
  { name = "Dein Name", email = "dein.email@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "pyosmium>=3.4.0",
  "geopandas>=0.12.0",
  "fiona>=1.8.0",
  "shapely>=2.0.0",
  "click>=8.0.0",
  "matplotlib>=3.5.0"
]

[project.scripts]
plotseamap = "plotseamap.cli:main"